Basic Player movement done
The Player can now use WASD to move around the world. There is a slight problem with this since it is done in the project settings which UE5 doesn't want us to use that method anymore but the other method is proving to have issues. Will need to come back to this at a later date.