Added Sprint, slide and jump
The player is now able to sprint when they press left shift along with the ability to jump and even double jump when the space bar is pressed. The last addition added was the ability to slide when sprinting. This is a simple implementation as it only create a temporary offset and moves the camera down and back up again to simulate sliding. Needs more work to change the players speed and move the collision of the player as well