Created the base blueprints for traps
This includes the hanging car which I need to add the ability for it to fall, as well as damage whatever's below.

The Car crusher, which will lerp between two positions, applying max damage to whatever is between it. It must loop this lerp after activated.

Finally, there's the button, which will be activated upon being shot, and will activate whatever shares the same id as it, meaning it can be used for the car crushers as well as doors or any number of things.