Base time slow cooldown implemented
The last commit that said I did cooldown, what I meant was that I gave the slow down effect a timer before it ends.

I now just need to make it so that if the slow down effect is ended early, then the cooldown will be shortened depending on the time it was cancelled.